apply black, add .flake8

+++ autoforecast/src/utils/utils.py
@@ -4,9 +4,9 @@ import pandas as pd
 from sklearn.preprocessing import MinMaxScaler
 
 
-def encode(data, col='bank'):
+def encode(data, col="bank"):
     map_col_to_col_id = {col: col_id for col_id, col in enumerate(data[col].unique())}
-    data[f'{col}_token'] = data[col].map(map_col_to_col_id)
+    data[f"{col}_token"] = data[col].map(map_col_to_col_id)
     return data, map_col_to_col_id
 
 
@@ -24,12 +24,12 @@ def scale_transform(train, test=None):
     return train, test, scaler
 
 
-def prepare_for_training(df, cols_to_drop=['heading_id', 'company_id']):
+def prepare_for_training(df, cols_to_drop=["heading_id", "company_id"]):
     df = df.drop(cols_to_drop, axis=1)
     df.date = pd.to_datetime(df.date)
     df = df.set_index("date")
-    if 'month' not in cols_to_drop:
-        df['month'] = df['month'].astype('float64')
+    if "month" not in cols_to_drop:
+        df["month"] = df["month"].astype("float64")
     return df
